Append Influence to a Dataframe on the server side

Appends influence to a specified column for a given set of ids and periods in a dataframe on the server side

ds.AppendInfluence(
  df,
  influences,
  id_period_vector,
  column,
  newobj = NULL,
  datasources = NULL
)

Arguments

seed:

Seed used for random sampling

df:

The dataframe to append influence to represented as a string

influences:

Vector of influences to append represented as a string

id_period_vector:

Vector of ids and periods to append influence to represented as a string

column:

The column in the dataframe to append the influence to represented as a string

newobj:

The name of the new object

datasources:

The Datashield connections to use