clusterInfluenceFunction

The function calculates the influence of a cluster of variables on a target variable using an influence matrix on the server side.lows for the option to assign the result to a new object or a specific data source.

ds.did(
  data = NULL,
  yname = NULL,
  tname = NULL,
  idname = NULL,
  gname = NULL,
  t_periods = NULL,
  g_periods = NULL,
  xformla = NULL,
  control_group = "notyettreated",
  base_period = "varying",
  anticipation = 0,
  alpha = 0.05,
  bstrap = FALSE,
  biters = 1000,
  cband = TRUE,
  clustervars = NULL,
  est_method = "dr",
  maxit = 10000,
  datasources = NULL,
  clear_console = FALSE
)

Arguments

data

A string name of the dataframe containing the data on the server side.

yname

Name of the outcome variable in the data frame on the server side.

tname

Name of the time variable in the data frame on the server side.

idname

Name of the id variable in the data frame on the server side.

gname

Name of the treatment indicator variable in the data frame on the server side.

t_periods

Vector of time periods.

g_periods

Vector of treatment periods.

xformla

Formula of covariates.

control_group

either "nevertreated" or "notyettreated"

base_period

either "varying" or "universal".

anticipation

integer of pre-treatment anticipation.

alpha

Confidence level of CIs.

bstrap

If TRUE, multiplier bootstrap is used to compute standard errors.

biters

Number of bootstrap draws. Only relevenat if bstrap is TRUE.

cband

if TRUE, simultaneous confidence bands are returned.

clustervars

A string name of the vector of variables that make up the cluster for standard errors. Only has an influence if bstrap is true.

est_method

can either be "dr", "reg" or "ipw".

maxit

MAximal number of iterations for ds.glm.

datasources

A specific Datashield data source to which the result should be assigned.

clear_console

If TRUE, the coinsole is yleared at certain stages.