Subset Dataframe

The function is used to subset a dataframe based on a specified variable and value. The subsetted dataframe is assigned to a new object on the server side.

ds.subsetDf(df, name_variable, value, include_zero, newobj, datasources = NULL)

Arguments

df

A string with the name of the dataframe from the server side to be subsetted.

name_variable

The name of the variable to subset on.

value

The value to subset the variable on.

include_zero

Logical. Should zero values be included in the subset?

newobj

The name of the new object to assign the subsetted dataframe to.

datasources

A specific Datashield data source to which the result should be assigned.